fix windwos bug

+++ lab_tools/spectrogram.py
@@ -256,6 +256,9 @@ def zip_directory_with_command(directory_path, output_zip_path):
         directory_path (str): 圧縮するディレクトリのパス。
         output_zip_path (str): 出力するzipファイルのパス。
     """
+
+    if os.name == 'nt':
+        return None
     try:
         # zipコマンドを実行
         subprocess.run(['zip', '-j', '-r', output_zip_path, directory_path], check=True)
@@ -358,7 +361,8 @@ def generate_spectrogram_and_signal_plot(
         config_yaml_path
     )
 
-    os.remove('/tmp/all_analyze_file.zip')
+    if os.path.exists('/tmp/all_analyze_file.zip'):
+        os.remove('/tmp/all_analyze_file.zip')
     ziped_file = zip_directory_with_command(output_dir, "/tmp/all_analyze_file.zip")
 
     return spectrogram_plot_path, plot_frequency_band_intensity_path, signal_plot_path, ziped_file
